src/use-cases/pets/get-pet-details.ts
import { Pet } from "@prisma/client";
import { OrgsRepository } from "../../repositories/orgs-repository";
import { PetsRepository } from "../../repositories/pets-repository";
import { ResourceNotFoundError } from "../errors/resource-not-found-error";

interface GetPetDetailsUseCaseRequest {
  petId: string
}

interface GetPetDetailsUseCaseResponse {
  details: {
    orgPhone: string
    pet: Pet
  }
}

export class GetPetDetailsUseCase {
  constructor(
    private orgsRepository: OrgsRepository,
    private petsRepository: PetsRepository
) {}

  async execute({
    petId,
  }: GetPetDetailsUseCaseRequest): Promise<GetPetDetailsUseCaseResponse> {
    const pet = await this.petsRepository.findPetById(petId)

    if (!pet) {
      throw new ResourceNotFoundError()
    }

    const org = await this.orgsRepository.findById(pet.org_id)

    if (!org) {
      throw new ResourceNotFoundError()
    }

    const orgPhone = org.phone

    return {
      details: {
        orgPhone, 
        pet,
      }
    };
  }
}
